Update token_metadata tests
